pyproject.toml
[project]
name = "scraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "openai>=2.8.1",
    "pillow>=11.0.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "streamlit<1.33",
    "tinydb>=4.8.2",
]
